requirements.txt
SmartApi==1.3.5
pyotp==2.9.0
logzero==1.7.0
python-dotenv==1.0.0
fastmcp==0.1.0
typing-extensions>=4.0.0 